requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.2.0
drf-yasg==1.21.7
filetype==1.2.0
geographiclib==2.0
geopy==2.4.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauth2-provider==0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
phonenumbers==8.13.35
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
